pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deer-rag"
version = "0.1.0"
description = "Modular RAG and retrieval optimization system focused on hybrid retrieval, token-efficient context assembly, and evaluation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
  "sqlalchemy>=2.0.0",
  "aiosqlite>=0.20.0",
  "pyyaml>=6.0.0",
  "tiktoken>=0.7.0",
  "numpy>=1.26.0",
  "faiss-cpu>=1.8.0",
  "rank-bm25>=0.2.2",
  "sentence-transformers>=3.0.0",
  "pymupdf>=1.24.0",
  "trafilatura>=1.9.0",
  "beautifulsoup4>=4.12.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*"]
